+++ src/mailman/handlers/tests/test_mimedel.py
@@ -46,11 +46,14 @@ from zope.component import getUtility
 def dummy_script(arg=''):
     exe = sys.executable
     non_ascii = ''
+    report = 'no'
     if arg == 'non-ascii':
         non_ascii = '‘...’'
     extra = ''
     if arg == 'scripterr':
         extra = 'error'
+    if arg == 'report':
+        report = 'yes'
     with ExitStack() as resources:
         tempdir = tempfile.mkdtemp()
         resources.callback(shutil.rmtree, tempdir)
@@ -69,7 +72,9 @@ print(open(sys.argv[1]).readlines()[0])
         config.push('dummy script', """\
 [mailman]
 html_to_plain_text_command = {exe} {script} {extra} $filename
-""".format(exe=exe, script=filter_path, extra=extra))
+filter_report = {report}
+""".format(exe=exe, script=filter_path, extra=extra, report=report))
+
         resources.callback(config.pop, 'dummy script')
         if arg == 'nonexist':
             os.rename(filter_path, filter_path + 'xxx')
@@ -533,4 +538,83 @@ spreadsheet
         self.assertEqual(msg['content-type'], 'text/plain; charset="utf-8"')
         self.assertEqual(msg.get_payload(decode=True), b"""\
 Plain text
+""")
+
+    def test_report(self):
+        # Hit all the pass and filter conditions for reporting
+        self._mlist.pass_extensions = ['txt']
+        self._mlist.filter_types = ['image']
+        self._mlist.pass_types = ['text', 'application', 'multipart']
+        msg = mfs("""\
+From: anne@example.com
+To: test@example.com
+Subject: Testing mixed extension
+Message-ID: <ant>
+MIME-Version: 1.0
+Content-Type: multipart/mixed; boundary="AAAA"
+
+--AAAA
+Content-Type: multipart/alternative; boundary="BBBB"
+
+--BBBB
+Content-Type: text/plain; charset="utf-8"
+
+Plain text
+
+--BBBB
+Content-Type: text/html; charset="utf-8"
+
+HTML text
+
+--BBBB--
+--AAAA
+Content-Type: application/octet-stream; name="test.xlsX"
+Content-Disposition: attachment; filename="test.xlsX"
+
+spreadsheet
+
+--AAAA
+Content-Type: application/octet-stream; name="test.exe"
+Content-Disposition: attachment; filename="test.exe"
+
+executable
+
+--AAAA
+Content-Type: image/jpeg; name="My_image"
+Content-Disposition: inline; filename="My_image"
+
+image
+
+--AAAA
+Content-Type: video/mp4; name="My_video"
+Content-Disposition: inline; filename="My_video"
+
+video
+
+--AAAA--
+""")
+        process = config.handlers['mime-delete'].process
+        with dummy_script('report'):
+            process(self._mlist, msg, {})
+        self.assertEqual(msg['content-type'], 'text/plain; charset="utf-8"')
+        self.assertEqual(msg.get_payload(decode=True), b"""\
+Plain text
+
+___________________________________________
+Mailman's content filtering has removed the
+following MIME parts from this message.
+
+Content-Type: application/octet-stream
+    Name: test.xlsX
+
+Content-Type: application/octet-stream
+    Name: test.exe
+
+Content-Type: image/jpeg
+    Name: My_image
+
+Content-Type: video/mp4
+    Name: My_video
+
+Replaced multipart/alternative part with first alternative.
 """)